+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqModifyTopicName.java
@@ -4,6 +4,6 @@ import org.codehaus.jackson.map.annotate.JsonSerialize;
 
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
 public class ReqModifyTopicName {
-    public int teamId;
+    public long teamId;
     public String name;
 }
