+++ app/src/main/java/com/tosslab/jandi/app/ui/maintab/topic/views/folderlist/adapter/DragnDropTouchHelper.java
@@ -12,7 +12,7 @@ public class DragnDropTouchHelper extends ItemTouchHelper.SimpleCallback {
 
     private TopicFolderSettingAdapter topicFolderSettingAdapter;
 
-    private int startFolderId;
+    private long startFolderId;
     private int dropPosition;
 
     public DragnDropTouchHelper(TopicFolderSettingAdapter topicFolderChooseAdapter) {
